Update to latest engine Fix imports of moved classes
