+++ pom.xml
@@ -49,8 +49,8 @@
     </modules>
 
     <properties>
-        <legend.engine.version>4.73.1</legend.engine.version>
-        <legend.sdlc.version>0.186.0</legend.sdlc.version>
+        <legend.engine.version>4.78.1</legend.engine.version>
+        <legend.sdlc.version>0.188.1</legend.sdlc.version>
 
         <logback.version>1.4.14</logback.version>
